Re-arranged project layout - testing for best global control of biquad array
