+++ Modules/stoneydsp_core/application/stoneydsp_ConsoleApplication.hpp
