+++ Modules/stoneydsp_core/maths/stoneydsp_Coefficient.cpp
