+++ Modules/stoneydsp_core/maths/stoneydsp_MathsConstants.hpp
