+++ Modules/stoneydsp_core/maths/stoneydsp_MathsFunctions.hpp
