+++ Modules/stoneydsp_core/maths/stoneydsp_MathsIConstants.hpp
