+++ Modules/stoneydsp_core/maths/stoneydsp_MathsIFunctions.hpp
