+++ Modules/stoneydsp_core/system/stoneydsp_StandardHeader.hpp
