+++ Sources/Processor.hpp
@@ -25,6 +25,21 @@
 #pragma once
 #define STONEYDSP_BIQUADS_PROCESSOR_HPP_INCLUDED
 
+#include <juce_audio_basics/juce_audio_basics.h>                                // depends: juce_core
+#include <juce_audio_formats/juce_audio_formats.h>                              // depends: juce_audio_basics
+#include <juce_audio_processors/juce_audio_processors.h>                        // depends: juce_gui_extra, juce_audio_basics
+#include <juce_core/juce_core.h>                                                //
+#include <juce_data_structures/juce_data_structures.h>                          // depends: juce_events
+#include <juce_dsp/juce_dsp.h>                                                  // depends: juce_audio_formats
+#include <juce_events/juce_events.h>                                            // depends: juce_core
+#include <juce_graphics/juce_graphics.h>                                        // depends: juce_events
+#include <juce_gui_basics/juce_gui_basics.h>                                    // depends: juce_graphics, juce_data_structures
+#include <juce_gui_extra/juce_gui_extra.h>                                      // depends: juce_gui_basics
+
+#include "Biquads.hpp"
+#include "Parameters.hpp"
+#include "Wrapper.hpp"
+
 namespace StoneyDSP {
 /** @addtogroup StoneyDSP @{ */
 
@@ -76,9 +91,9 @@ public:
     void getCurrentProgramStateInformation(juce::MemoryBlock& destData) override;
     void setStateInformation (const void* data, int sizeInBytes) override;
     void setCurrentProgramStateInformation(const void* data, int sizeInBytes) override;
-    // //==============================================================================
+    //==============================================================================
     juce::UndoManager& getUndoManager() { return undoManager; }
-    // //==============================================================================
+    //==============================================================================
     juce::AudioProcessorValueTreeState& getApvts() { return apvts; }
     //==============================================================================
     juce::dsp::ProcessSpec& getSpec() { return spec; }
